Feature: EntityAwareGameComponent, change Transform to use this
This interface was created to prevent the need for passing in the entity
that a component is being added to. Instead, GameEntities will automatically
register themselves with EntityAwareGameComponents - even if that component
is moved to another GameEntity. This means components can now be
aware of the entity they are part of just by implementing the interface