Change RenderOrder to order from bottom left, fix ItemsRowEntity for oversized items
This commit aims to simplify the offsetting system so the faux depth can be achieved
with less manual tweaking. The RenderSystem now orders objects from smallest to largest
bottom left value instead of top left. This commit also features a rewrite of ItemsRowEntity.
All positions on a farm are now represented by an individual entity with a FarmSpriteComponent.
This is controlled by FarmSpriteSystem, which tracks the growth and selects the correct sprite
accordingly. The goal of this was for oversized items to order correctly.